refactor: migrate TicketConsumer and RabbitMqConsumerBase to new namespace structure
